// apps/studio/src/api/model/requestPresignedUploadDtoType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * FocusFlow Studio API
 * FocusFlow Mode-B SaaS Enterprise API Contract (statically generated via TypeScript AST and DTO reflection)
 * OpenAPI spec version: 1.0.0
 */

/**
 * Category of asset being uploaded
 */
export type RequestPresignedUploadDtoType = typeof RequestPresignedUploadDtoType[keyof typeof RequestPresignedUploadDtoType];


export const RequestPresignedUploadDtoType = {
  BACKGROUND_IMAGE: 'BACKGROUND_IMAGE',
  OVERLAY_IMAGE: 'OVERLAY_IMAGE',
  VOICEOVER_AUDIO: 'VOICEOVER_AUDIO',
  BGM_AUDIO: 'BGM_AUDIO',
} as const;